feat(ui): NNotification component
