+++ packages/devtools/client/composables/dialog.ts
@@ -12,18 +12,3 @@ interface RestartDialog {
 export function useRestartDialogs() {
   return useState<RestartDialog[]>('devtools:restart-dialogs', () => [])
 }
-
-let _showNotification: typeof showNotification
-
-export function showNotification(data: {
-  message: string
-  icon?: string
-  classes?: string
-  duration?: number
-}) {
-  _showNotification?.(data)
-}
-
-export function provideNotificationFn(fn: typeof showNotification) {
-  _showNotification = fn
-}
